small bug fixes

+++ src/util/macroDict.c
@@ -5,6 +5,7 @@
 #include <limits.h>
 #include <math.h>
 #include <time.h>
+#include <stdint.h>
 
 #include "macroDict.h"
 #include "callbacks.h"
@@ -55,7 +56,7 @@ void _insertLine(TokList *list, Token const *tok) {
 	Token new;
 	cpToken(&new, tok);
 	new.type = TT_NUMB_CONSTANT;
-	int strLength = log10(tok->posLine) + 2;
+	int strLength = (int) log10((double) tok->posLine) + 2;
 	new.contents = malloc(strLength * sizeof(char));
 	snprintf(new.contents, strLength, "%d", new.posLine);
